Restcture files
