+++ rask-engine/src/resources/resource_table.rs
@@ -2,7 +2,7 @@ use super::Resource;
 use crate::EngineError;
 
 /// The library is used to store and retrieve resources.
-pub struct ResourceTable(&'static mut [Resource]);
+pub struct ResourceTable([Resource; super::RESOURCE_COUNT as usize]);
 
 macro_rules! character_check_helper {
     (Texture, $value: ident) => {
