+++ wasm/src/communication/message_queue.rs
