+++ wasm/src/graphics/mod.rs
@@ -1,0 +1,6 @@
+pub mod context;
+mod render;
+// mod shader
+mod graphics_api;
+#[doc(inline)]
+pub use graphics_api::GraphicsApi;
